Alcohol and Tobacco Expenditures in Belgian Households

Data on budget shares of alcohol and tobacco for 2724 Belgian households, taken from the Belgian household budget survey of 1995/96.

data("AlcoholTobacco", package = "htobit")

Format

A data frame containing 2724 observations on 9 variables.

alcohol

Budget share of alcohol.

tobacco

Budget share of tobacco.

expenditure

Total expenditure.

age

10-year age intervals ranging from 0 (younger than 30) to 4 (60 or older).

adults

Number of adults.

oldkids

Number of old kids (2 years or older).

youngkids

Number of young kinds (younger than 2 years).

occupation

Factor indicating occupation of household head ("blue" collar worker, "white" collar worker, or "other").

region

Factor indicating region of residence (Dutch-speaking "flanders", French-speaking "wallonie", or bilingual "brussels").

Source

Supplementary material for Verbeek (2004), originally provided by the National Institute of Statistics (NIS), Belgium.

References

Verbeek, M. (2004). A Guide to Modern Econometrics, 2nd ed. Chichester, UK: John Wiley.

Examples

data("AlcoholTobacco", package = "htobit")

## homoscedastic tobit model for budget share of alcohol
ma <- htobit(alcohol ~ (age + adults) * log(expenditure) + oldkids + youngkids,
  data = AlcoholTobacco)
summary(ma)

#> 
#> Call:
#> htobit(formula = alcohol ~ (age + adults) * log(expenditure) + oldkids + 
#>     youngkids, data = AlcoholTobacco)
#> 
#> Standardized residuals:
#>     Min      1Q  Median      3Q     Max 
#> -1.0698 -0.4407 -0.1364  0.3934  8.3170 
#> 
#> Coefficients (location model):
#>                           Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)             -0.1591533  0.0437782  -3.635 0.000278 ***
#> age                      0.0134938  0.0108824   1.240 0.214989    
#> adults                   0.0291901  0.0169469   1.722 0.084989 .  
#> log(expenditure)         0.0126679  0.0032156   3.939 8.17e-05 ***
#> oldkids                 -0.0026408  0.0006049  -4.366 1.27e-05 ***
#> youngkids               -0.0038789  0.0023835  -1.627 0.103651    
#> age:log(expenditure)    -0.0008093  0.0008006  -1.011 0.312067    
#> adults:log(expenditure) -0.0022484  0.0012232  -1.838 0.066051 .  
#> 
#> Coefficients (scale model with log link):
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -3.71236    0.01533  -242.1   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 
#> Log-likelihood:  4755 on 9 Df
#> Number of iterations in BFGS optimization: 16 

## homoscedastic tobit model for budget share of tobacco
mt <- htobit(tobacco ~ (age + adults) * log(expenditure) + oldkids + youngkids,
  data = AlcoholTobacco)
summary(mt)

#> 
#> Call:
#> htobit(formula = tobacco ~ (age + adults) * log(expenditure) + oldkids + 
#>     youngkids, data = AlcoholTobacco)
#> 
#> Standardized residuals:
#>     Min      1Q  Median      3Q     Max 
#> -0.9993  0.3364  0.5176  0.7105  4.2755 
#> 
#> Coefficients (location model):
#>                           Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)              0.5899799  0.0930472   6.341 2.29e-10 ***
#> age                     -0.1258528  0.0241330  -5.215 1.84e-07 ***
#> adults                   0.0153699  0.0379745   0.405  0.68567    
#> log(expenditure)        -0.0444314  0.0068613  -6.476 9.44e-11 ***
#> oldkids                  0.0042697  0.0013242   3.224  0.00126 ** 
#> youngkids               -0.0099719  0.0054708  -1.823  0.06834 .  
#> age:log(expenditure)     0.0088221  0.0017798   4.957 7.17e-07 ***
#> adults:log(expenditure) -0.0006007  0.0027448  -0.219  0.82675    
#> 
#> Coefficients (scale model with log link):
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -3.03666    0.02465  -123.2   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 
#> Log-likelihood: 758.7 on 9 Df
#> Number of iterations in BFGS optimization: 31